Database Fundamentals Challenge
The challenge is to follow the steps of the database development process, starting with conceptual modeling with the creation of an Entity-Relationship Diagram (DER). Then, it will be necessary to perform the logical modeling using the Relational Model, apply normalization to ensure data integrity, create the tables according to the defined logical model and insert data directly into the chosen database.

In addition, the challenge includes defining functions in relational algebra to extract information from the database and demonstrate the effectiveness of the modeled structure.

With this challenge, it seeks to provide opportunities for the application of the knowledge acquired throughout the activity, from modeling to practical implementation, putting into practice the skills of creating and manipulating databases.